Bugfix for excluding nested keys
If a tensordict contains nested keys such as ("nested", "a") and
("nested", "b"), then `tensordict.exclude("nested")` was only removing
"nested" from the output of `tensordict.keys(include_nested=True)` and
then selecting on the results of that. But that output would of course
include `("nested", "a")` and `("nested", "b")`, so the tensordict would
not get filtered as intended.